refactor(desktop): split boot-time event sync out of migration.rs
migration.rs crossed the file-size guard (1577 > 1575) when the
managed-agent reconcile call site landed. Rather than bump the limit,
extract the post-identity event-sync leg (run_event_sync +
migrate_personas_to_events/_in_dir + migrate_teams_to_events/_in_dir)
into a new event_sync.rs with its tests, and ratchet the guard down:
migration.rs 1575 -> 1310, migration_tests.rs 1410 -> 1110.

Pure code move — no behavior change; the full test set is unchanged
(verified by name against the pre-split list).
